Added basic p2pfl examples
