+++ examples/node1.py
@@ -1,0 +1,36 @@
+# 
+# This file is part of the federated_learning_p2p (p2pfl) distribution (see https://github.com/pguijas/federated_learning_p2p).
+# Copyright (c) 2022 Pedro Guijas Bravo.
+# 
+# This program is free software: you can redistribute it and/or modify  
+# it under the terms of the GNU General Public License as published by  
+# the Free Software Foundation, version 3.
+#
+# This program is distributed in the hope that it will be useful, but 
+# WITHOUT ANY WARRANTY; without even the implied warranty of 
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU 
+# General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License 
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
+#
+
+from p2pfl.node import Node
+from p2pfl.learning.pytorch.mnist_examples.mnistfederated_dm import MnistFederatedDM
+from p2pfl.learning.pytorch.mnist_examples.models.mlp import MLP
+import sys
+
+if __name__ == "__main__":
+
+    if len(sys.argv) != 2:
+        print("Usage: python3 nodo1.py <self_port>")
+        sys.exit(1)
+
+    node = Node(
+        MLP(),
+        MnistFederatedDM(sub_id=0, number_sub=2),
+        port=int(sys.argv[1]),
+    )
+    node.start()
+    
+    input("Press any key to stop\n")
